Fixed small issue in r2046 where last character of a field wasn't updated on backspace/delete.
........
create_win_dist.sh:
* cater for TMPDIR (including TMP/TEMP) set by user
* echo intermediate steps
* fixed out-of-tree "make dist"

build_windows:
* defaults.h.in:
  * changed install folder for DIST builds to PACKAGE_NAME + _ + PACKAGE_VERSION (GnuCOBOL_2.3_dev)
  * use lib / lib_x64 like created in makedist.bat
* makedist.bat:
  * use PACKAGE_NAME, PACKAGE_VERSION and suffix "vs_bin" for resulting 7zip file, fallback to directory, when 7zip is not available
  * fixed version lookup when directory contains spaces
* README.txt: updated
* set_env_vs.bat, set_env_vs_x64.bat, set_env_vs_dist.bat, set_env_vs_dist_x64.bat: only try to setup a matching cl.exe if it isn't present already
........
cobc:
* typeck.c (cb_validate_program_data): temporarily disabled ASSIGN fields with changing address (BASED + LINKAGE) to prevent SIGSEGV until code generation is changed, see [bugs:#421]
........
Suppress runtime error on Windows (bug #420).
........
Follow-up to r1763 - cobc:
* reserved.c: missing adjustments of using cb_feature_mode in intrinsics
........
Major change to NEWS for 2.2 Release
........
cobc:
* codegen.c: changed generated newlines/comments for file initialization
........
build_windows: version number bump
........
po:
* update_linguas.sh: always include the generated en@boldquot en@quot in LINGUAS on recreation
* general: updated translations from TP
* gnucobol.pot: recreated by running `make update-po`
........


git-svn-id: https://svn.code.sf.net/p/open-cobol/code/branches/gnucobol-2.x@2060 ed166372-6744-4ac0-a67f-bb1ae9efa102

+++ create_win_dist.sh
@@ -44,26 +44,47 @@ if test ! -d "$EXTSRCDIR/build_windows"; then
 	exit 5
 fi
 
+# check where to place temporary files
+if test "x$TMPDIR" = "x"; then
+	if test "x$TMP" = "x"; then
+		if test "x$TEMP" = "x"; then
+			TMPDIR=/tmp
+		else
+			TMPDIR=$TEMP
+		fi
+	else
+		TMPDIR=$TMP
+	fi
+fi
+if test ! -d "$TMPDIR"; then
+	echo "TMPDIR $TMPDIR" does not exist, aborting $0
+	exit 1
+fi
+export TMPDIR
 
 # Create temporary folder as we don't want to change the EXTDISTDIR's content
-WINTMP=/tmp/win-dist-$(date +%s)
+WINTMP=$TMPDIR/win-dist-$(date +%s)
 
 rm -r -f $WINTMP
+echo mkdir $WINTMP
 mkdir $WINTMP
 
-cp -p -r --parents $EXTDISTDIR $WINTMP || exit 1
+echo cp -p -r  $EXTDISTDIR $WINTMP
+cp -p -r $EXTDISTDIR $WINTMP || exit 1
 
 # Add content only necessary for windows dist zip
-cp -p -r --parents $EXTSRCDIR/build_windows $WINTMP/$EXTDISTDIR/ || exit 2
+echo cp -p -r $EXTSRCDIR/build_windows $WINTMP/$EXTDISTDIR/
+cp -p -r $EXTSRCDIR/build_windows $WINTMP/$EXTDISTDIR/ || exit 2
+echo cp $EXTSRCDIR/tests/atlocal_win $WINTMP/$EXTDISTDIR/tests/atlocal_win
 cp $EXTSRCDIR/tests/atlocal_win $WINTMP/$EXTDISTDIR/tests/atlocal_win || exit 2
 
 olddir=$(pwd)
 cd $WINTMP/$EXTDISTDIR || exit 3
 
 # rename templates for faster setup
-cd build_windows
-mv config.h.in config.h
-mv defaults.h.in defaults.h
+cd build_windows || exit 5
+mv "config.h.in"   "config.h"
+mv "defaults.h.in" "defaults.h"
 cd ..
 
 # Remove content not necessary for windows dist zip --> breaks make dist[check]
@@ -76,16 +97,16 @@ find -regextype posix-egrep -regex ".*(\.([chyl]|def|cpy|cob|conf|cfg)|(README|C
  
 # fix timestamps again
 ./doc/cobcinfo.sh "fixtimestamps"
-touch ./bin/cobcrun.1
-touch ./cobc/cobc.1
-touch ./cobc/ppparse.c
-touch ./cobc/parser.c
-#touch ./cobc/pplex.c
-#touch ./cobc/scanner.c
-#touch ./libcob/libcob.3
+touch "./bin/cobcrun.1"
+touch "./cobc/cobc.1"
+touch "./cobc/ppparse.c"
+touch "./cobc/parser.c"
+#touch "./cobc/pplex.c"
+#touch "./cobc/scanner.c"
+#touch "./libcob/libcob.3"
 
 # bugfix for old _MSC versions that define __STDC_VERSION__ >= 199901L but don't work correct
-for file in ./cobc/pplex.c ./cobc/scanner.c; do
+for file in "./cobc/pplex.c" "./cobc/scanner.c"; do
 # "sed -i" isn't supported on all systems --> maybe use sed && mv - do we actually want to care for this here?
 #	sed -e 's/199901L/199901L \&\&(!defined(_MSC_VER) || _MSC_VER >= 1800)/g' \
 #	  $file > $file.tmp && mv -f $file.tmp $file
